mi-proyecto/backend/requirements.txt
flask==3.0.2
flask-jwt-extended==4.6.0
flask-limiter==3.5.0
flask-cors==4.0.0
bcrypt==4.1.2
python-dotenv==1.0.1
bleach==6.1.0
